#102: Corrigidas mensagens de erro.
